Move files into api and jobs folder for better organization
